Preprocess images
Extract features from each image in a given directory using Xception
model. The model is slightly modified as the last layer is removed to
retrieve the extracted features instead of a prediction.

# Beware: this is experimental as the paper about merge models described
by Marc Tanti, et al. in their 2017 papers is a VGG19. Results could be
not as accurate as expected.

# References:
- Where to put the Image in an Image Caption Generator, 2017.
- What is the Role of Recurrent Neural Networks (RNNs) in an Image
Caption Generator?, 2017.